+++ packages/gator-permissions-controller/CHANGELOG.md
@@ -10,5 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial release ([#6033](https://github.com/MetaMask/core/pull/6033))
+- Integrate erc7715 types from @metamask/7715-permission-types package ([#6379](https://github.com/MetaMask/core/pull/6379))
 
 [Unreleased]: https://github.com/MetaMask/core/